feat: schedule update and delete endpoints (#36)

+++ AurorumClinic/src/main/java/pl/edu/pja/aurorumclinic/features/appointments/employees/commands/DeleteAppointment.java
@@ -13,7 +13,11 @@ import pl.edu.pja.aurorumclinic.shared.data.AppointmentRepository;
 import pl.edu.pja.aurorumclinic.features.appointments.shared.events.AppointmentDeletedEvent;
 import pl.edu.pja.aurorumclinic.shared.ApiResponse;
 import pl.edu.pja.aurorumclinic.shared.data.models.Appointment;
+import pl.edu.pja.aurorumclinic.shared.data.models.enums.AppointmentStatus;
+import pl.edu.pja.aurorumclinic.shared.exceptions.ApiException;
 import pl.edu.pja.aurorumclinic.shared.exceptions.ApiNotFoundException;
+
+import java.util.Objects;
 
 @RestController
 @RequestMapping("/api/appointments")
@@ -34,6 +38,9 @@ public class DeleteAppointment {
     private void handle(Long appointmentId) {
         Appointment appointmentFromDb = appointmentRepository.findById(appointmentId)
                 .orElseThrow(() -> new ApiNotFoundException("Id not found", "id"));
+        if (Objects.equals(appointmentFromDb.getStatus(),AppointmentStatus.FINISHED)) {
+            throw new ApiException("Appointment has finished, unable to delete", "id");
+        }
         appointmentRepository.delete(appointmentFromDb);
         applicationEventPublisher.publishEvent(
                 new AppointmentDeletedEvent(appointmentFromDb.getPatient(), appointmentFromDb));
